Added the corresponding dataset
In this commit, I have added the dataset namely "wifi_localization.txt" on which the K means algorithm is running. 
This dataset was collected to perform experimentation on how WIFI signal strengths can be used to determine one of the indoor locations. The target variable belongs to (1-4). In this dataset each attribute is WIFI signal strength observed on smartphone. There are 7 different sample of WIFI strength. And a single target variable